requirements.txt
gym==0.23.1 # Need to do atari, mujoco afterwards for pip reqsons
absl-py
atari_py
# They made a breaking change in a future dopamine version
# Here, I think. https://github.com/google/dopamine/tree/55ec7925cb944869127ec4755544caa6bbcdc37e
dopamine-rl==4.0.0
gin-config >= 0.1
matplotlib
numpy >= 1.14,<1.24
pillow
pygame
six
scipy
tensorflow<2.8.0,>=2.7.0
tensorflow-probability<0.16,>=0.15.0
autorom[accept-rom-license]
cpprb
seeding

lockfile # sometimes needed for successful mujoco install
git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl
protobuf==3.19.4 # d4rl changes protobuf value to something invalid for other packages

git+https://github.com/samlobel/onager@gin_multiworker

gym-robotics==0.1.0 # Possibly we need --no-reqs. Not sure if we work with that gym version...

jax==0.3.20
jaxlib==0.3.20
flax==0.6.1
optax==0.1.3
chex==0.1.5

seaborn
# Needs --no-deps and --ignore-requires-python
# git+https://github.com/camall3n/visgrid@6f7f3a6373e478dbc64e27a692d75f499e5870e0
